requirements.txt
Flask
pyodbc
python-magic
pillow